requirements/requirements.retrain.txt
numpy==1.23.5 
pandas==1.5.2 
scikit-learn==1.2.0 
pytest
PyYAML
mlflow
xgboost
